feat(core): Add Gossiper helper
This helper provides a class that can be used to implement simple gossip
behavior, where a node is picked at random at a certain interval. This
can be used to broadcast certain messages, such as requests or updates
of data in a way where it eventually reaches the entire group.